requirements.txt
aniso8601==9.0.1
attrs==23.2.0
bitarray==2.9.2
bitstring==4.2.3
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.1
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
esptool==4.7.0
fancycompleter==0.9.1
Flask==3.1.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
fonttools==4.51.0
idna==3.10
iniconfig==2.0.0
intelhex==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib==3.8.4
mirakuru==2.5.3
numpy==1.26.4
packaging==24.0
pdbpp==0.10.3
perlin-noise==1.12
pillow==10.3.0
pluggy==1.5.0
port-for==0.7.4
psutil==6.1.0
pycparser==2.22
pygame==2.5.2
Pygments==2.18.0
pymongo==4.10.1
pyodbc==5.2.0
PyOpenGL @ file:///C:/Users/paobb/Downloads/PyOpenGL-3.1.7-cp311-cp311-win_amd64.whl#sha256=97ae2903117797fbaae9776bcf3a78617a41088146c1543a72ed5e3ba2a5f3e5
PyOpenGL-accelerate @ file:///C:/Users/paobb/Downloads/PyOpenGL_accelerate-3.1.7-cp311-cp311-win_amd64.whl#sha256=b1f0d91b81e26ddf443841b12c27d86d6bd663eb8ef0b4e7e28fc859c5c75ae9
pyparsing==3.1.2
pyreadline==2.1
pyrepl==0.9.0
pyserial==3.5
pytest==8.3.3
pytest-mongo==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
reedsolo==1.7.0
requests==2.32.3
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.3
wmctrl==0.5
